cerebric_core/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cerebric-core"
version = "0.1.0"
description = "Cerebric Phase 1 runtime skeleton (agents, tools, ingestion, index, obs, eval)"
authors = [{ name = "Cerebric Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # Logging/typing
  "pydantic>=2.6",
  # Index (optional placeholder; actual wiring later)
  "chromadb>=0.5; python_version >= '3.10'",
  # Config parsing & watching
  "pyyaml>=6.0",
  "watchdog>=4.0",
  # JSON schema validation
  "jsonschema>=4.19",
  # Journald native reader (optional, falls back to journalctl)
  "systemd-python>=235; platform_system == 'Linux'",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["cerebric_core*"]
